Move TalentTemplateInfo from UnitCodes to MiscCodes, apply talent auras each time a talent is trained
