+++ database/world/WorldDatabaseManager.py
@@ -7,8 +7,7 @@ from sqlalchemy.orm import sessionmaker, scoped_session
 
 from database.world.WorldModels import *
 from utils.ConfigManager import *
-from utils.constants.MiscCodes import HighGuid
-from utils.constants.UnitCodes import TalentTemplateInfo
+from utils.constants.MiscCodes import HighGuid, TalentTemplateInfo
 
 DB_USER = os.getenv('MYSQL_USERNAME', config.Database.Connection.username)
 DB_PASSWORD = os.getenv('MYSQL_PASSWORD', config.Database.Connection.password)
